% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covWeight.R
\name{covWeight}
\alias{covWeight}
\title{Create a weighted covariance matrix}
\usage{
covWeight(m, weights, ...)
}
\arguments{
\item{m}{A \code{matrix}, e.g. as returned by \code{\link[terra:values]{terra::values()}}.}

\item{weights}{A \code{numeric} vector of weights. For lat/lon data this can be
produced with \code{\link[=getWeights]{getWeights()}}.}

\item{...}{Additional arguments passed to \code{\link[stats:cov.wt]{stats::cov.wt()}}}
}
\value{
See \code{\link[stats:cov.wt]{stats::cov.wt()}}.
}
\description{
Create a weighted covariance matrix
}
