% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{denoise}
\alias{denoise}
\title{Noise filtering through principal components}
\usage{
denoise(
  x,
  k = NULL,
  expl.var = NULL,
  weighted = TRUE,
  use.cpp = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{SpatRaster} (or \verb{Raster*}) series to be filtered.}

\item{k}{The number of components to be kept for reconstruction (ignored if
'expl.var' is supplied).}

\item{expl.var}{Minimum amount of variance to be kept after reconstruction
(should be set to \code{NULL} or omitted if 'k' is supplied).}

\item{weighted}{logical. If \code{TRUE} the covariance matrix will be
geographically weighted using the cosine of latitude during decomposition
(only important for lat/lon data).}

\item{use.cpp}{logical. Determines whether to use \strong{Rcpp} functionality,
defaults to \code{TRUE}.}

\item{verbose}{logical. If \code{TRUE} some details about the calculation process
will be output to the console.}

\item{...}{Additional arguments passed to \code{\link[stats:princomp]{stats::princomp()}}.}
}
\value{
A denoised \code{SpatRaster} series.
}
\description{
Filter noise from a raster series by decomposing into principal components
and subsequent reconstruction using only a subset of components.
}
\note{
Either 'k' or 'expl.var' must be specified. If both are supplied, 'k' will be
ignored. If none are supplied, an error will be thrown.
}
\examples{
gph = terra::unwrap(vdendool)
gph_dns = denoise(gph, expl.var = 0.8)

opar = par(mfrow = c(1,2))
plot(gph[[1]], main = "original")
plot(gph_dns[[1]], main = "denoised")
par(opar)

}
\seealso{
\code{\link[=anomalize]{anomalize()}}, \code{\link[=deseason]{deseason()}}
}
