% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deseasonC.R
\name{deseasonC}
\alias{deseasonC}
\title{Create seasonal anomalies (based on \strong{Rcpp})}
\usage{
deseasonC(x, cycle.window = 12, ...)
}
\arguments{
\item{x}{a RasterStack}

\item{cycle.window}{the window for the creation of the anomalies}

\item{...}{currently not used}
}
\value{
A deseasoned 'RasterStack'.
}
\description{
The function calculates seasonal anomalies from a 'RasterStack' object by
supplying a suitable seasonal window. To create monthly anomalies of a
'RasterStack' encompassing 12 layers per year, for instance, use
\code{cycle.window = 12}.
}
\examples{
data("australiaGPCP")

aus_dsnC <- deseasonC(australiaGPCP, 12)

opar <- par(mfrow = c(1,2))
plot(australiaGPCP[[1]], main = "original")
plot(aus_dsnC[[1]], main = "deseasoned")
par(opar)
}
\author{
Florian Detsch
}
\seealso{
\code{\link{anomalize}}, \code{\link{denoise}}
}

