% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{eot-class}
\alias{eot-class}
\title{Class definitions for remote}
\description{
Class definitions for remote
}
\section{Slots}{

\describe{
\item{\code{eot.series}}{the EOT time series at the identified base point}

\item{\code{max.xy}}{the cell number of the indeified base point}

\item{\code{exp.var}}{the (cumulative) explained variance of the considered EOT}

\item{\code{loc.eot}}{the location of the base point (in original coordinates)}

\item{\code{r.predictor}}{the RasterLayer of the correlation coefficients between the base point and each pixel of the predictor domain}

\item{\code{rsq.predictor}}{as above but for the coefficient of determination of the predictor domain}

\item{\code{rsq.sums.predictor}}{as above but for the sums of coefficient of determination of the predictor domain}

\item{\code{int.predictor}}{the RasterLayer of the intercept of the regression equation for each pixel of the predictor domain}

\item{\code{slp.predictor}}{same as above but for the slope of the regression equation for each pixel of the predictor domain}

\item{\code{p.predictor}}{the RasterLayer of the significance (p-value) of the the regression equation for each pixel of the predictor domain}

\item{\code{resid.predictor}}{the RasterBrick of the reduced data for the predictor domain}

\item{\code{r.response}}{the RasterLayer of the correlation coefficients between the base point and each pixel of the response domain}

\item{\code{rsq.response}}{as above but for the coefficient of determination of the response domain}

\item{\code{int.response}}{the RasterLayer of the intercept of the regression equation for each pixel of the response domain}

\item{\code{slp.response}}{as above but for the slope of the regression equation for each pixel of the response domain}

\item{\code{p.response}}{same the RasterLayer of the significance (p-value) of the the regression equation for each pixel of the response domain}

\item{\code{resid.response}}{the RasterBrick of the reduced data for the response domain}
}}

