% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAremoteClasses.R
\docType{class}
\name{EotStack-class}
\alias{EotStack-class}
\title{Class EotStack}
\description{
Class EotStack
}
\section{Slots}{

\describe{
\item{\code{modes}}{a list containing the individual 'EotMode's of the 'EotStack'}

\item{\code{names}}{the names of the modes}
}}

