% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names}
\alias{names}
\alias{names,EotStack-method}
\alias{names<-}
\alias{names<-,EotStack-method}
\alias{names,EotMode-method}
\alias{names<-,EotMode-method}
\title{Names of Eot* objects}
\usage{
\S4method{names}{EotStack}(x)

\S4method{names}{EotStack}(x) <- value

\S4method{names}{EotMode}(x)

\S4method{names}{EotMode}(x) <- value
}
\arguments{
\item{x}{a EotMode or EotStack}

\item{value}{name to be assigned}
}
\value{
if \code{x} is a EotStack, the names of all mdoes,
if \code{x} is a EotMode, the name the respective mode
}
\description{
Get or set names of Eot* objects
}
\examples{
gph <- terra::unwrap(vdendool)

nh_modes <- eot(gph, n = 2)

## mode names
names(nh_modes)
names(nh_modes) <- c("vdendool1", "vdendool2")

names(nh_modes)
names(nh_modes[[2]])

}
