% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmodes.R
\name{nmodes}
\alias{nmodes}
\title{Number of modes of an \code{EotStack}}
\usage{
nmodes(x)
}
\arguments{
\item{x}{An \code{EotStack}.}
}
\value{
The number of modes as \code{integer}.
}
\description{
Retrieves the number of modes of an \code{EotStack}.
}
\examples{
gph = terra::unwrap(vdendool)
nh_modes = eot(gph, n = 2)
nmodes(nh_modes)

}
