% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nXplain.R
\name{nXplain}
\alias{nXplain}
\alias{nXplain,EotStack-method}
\title{Number of EOTs needed for variance explanation}
\usage{
\S4method{nXplain}{EotStack}(x, var = 0.9)
}
\arguments{
\item{x}{An \code{EotStack}.}

\item{var}{The minimum amount of variance to be explained by the modes as
\code{numeric}, defaults to \code{0.9}.}
}
\value{
The number of EOTs needed to explain 'var'.
}
\description{
The function identifies the number of modes needed to explain a certain
amount of variance within the response field.
}
\note{
This is a post-hoc function. It needs an \code{EotStack} created as returned
by \code{\link[=eot]{eot()}}. Depending on the potency of the identified EOTs, it may be
necessary to compute a high number of modes in order to be able to explain
a large enough part of the variance.
}
\examples{
gph <- terra::unwrap(vdendool)

nh_modes <- eot(x = gph, y = NULL, n = 3, 
                standardised = FALSE, 
                verbose = TRUE)
             
### How many modes are needed to explain 25\% of variance?              
nXplain(nh_modes, 0.25)

}
