% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.tunnel.R
\name{open_tunnel}
\alias{open_tunnel}
\title{Open an SSH tunnel, prompting for passwords as needed}
\usage{
open_tunnel(remote_host, user = NULL, password = NULL,
  tunnel_dir = "~/.pecan/tunnel/", wait.time = 15,
  tunnel_script = "~/pecan/web/sshtunnel.sh")
}
\arguments{
\item{remote_host}{name of remote server to connect to (e.g. geo.bu.edu)}

\item{user}{username on remote_host}

\item{password}{password on remote_host}

\item{tunnel_dir}{directory to store tunnel file in, typically from settings$host}

\item{wait.time}{how long to give system to connect before deleting password (seconds)}

\item{tunnel_script}{Path to sshtunnel.sh script file for opening tunnel}
}
\value{
numeric giving ssh PID if configured, otherwise logical with TRUE = success
}
\description{
Open an SSH tunnel, prompting for passwords as needed
}
