% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-package.R
\docType{data}
\name{pacificSST}
\alias{pacificSST}
\title{Monthly SSTs for the tropical Pacific Ocean}
\format{
A \code{PackedSpatRaster} with the following attributes:\cr

\if{html}{\out{<div class="sourceCode sh">}}\preformatted{size        : 30, 140, 348  (nrow, ncol, nlyr)
resolution  : 1, 1  (x, y)
extent      : 150, 290, -15, 15  (xmin, xmax, ymin, ymax)
coord. ref. : +proj=longlat +ellps=WGS84 +towgs84=0,0,0,0,0,0,0 +no_defs
}\if{html}{\out{</div>}}
}
\usage{
pacificSST
}
\description{
Monthly NOAA sea surface temperatures for the tropical Pacific Ocean from 1982/01 to 2010/12
}
\details{
Monthly NOAA sea surface temperatures for the tropical Pacific Ocean from 1982/01 to 2010/12
}
\examples{
terra::unwrap(pacificSST)
}
\references{
Daily High-Resolution-Blended Analyses for Sea Surface Temperature\cr
Reynolds et al. (2007)\cr
Journal of Climate, Volume 20, Issue 22, pp. 5473 - 5496\cr
\doi{10.1175/2007JCLI1824.1}
}
