% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_get_jobid.R
\name{qsub_get_jobid}
\alias{qsub_get_jobid}
\title{Get Job ID from qsub output}
\usage{
qsub_get_jobid(out, qsub.jobid, stop.on.error)
}
\arguments{
\item{out}{Output from model execution, as a character.}

\item{qsub.jobid}{(character) Regular expression string for extracting job ID from qsub output.
Usually from \code{settings$host$qsub.jobid}}

\item{stop.on.error}{Throw error if \emph{any} of the runs fails. Default TRUE.}
}
\value{
Job ID, as a string
}
\description{
Get Job ID from qsub output
}
