% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_qsub_status.R
\name{qsub_run_finished}
\alias{qsub_run_finished}
\title{Check if qsub run finished}
\usage{
qsub_run_finished(run, host, qstat)
}
\arguments{
\item{run}{run ID, as an integer}

\item{host}{host structure to execute command on}

\item{qstat}{(string) qstat command for checking job status}
}
\value{
\code{TRUE} if run is marked as DONE, otherwise FALSE.
}
\description{
Check if qsub run finished
}
