% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbitmq.R
\name{rabbitmq_create_queue}
\alias{rabbitmq_create_queue}
\title{Create a queue in RabbitMQ.}
\usage{
rabbitmq_create_queue(
  url,
  auth,
  vhost,
  queue,
  auto_delete = FALSE,
  durable = TRUE
)
}
\arguments{
\item{url}{parsed RabbitMQ URL.}

\item{auth}{the httr authentication object to use.}

\item{vhost}{the vhost where to create the queue.}

\item{queue}{the queue that should be checked/created.}

\item{auto_delete}{should the queue be deleted afterwards (FALSE is default)}

\item{durable}{should the messages exists after a server restart (TRUE is default)}
}
\value{
TRUE if the queue now exists, FALSE otherwise.
}
\description{
This will first check to see if the queue already exists in RabbitMQ, if not
it will create the queue. If the queue exists, or is created it will return
TRUE, it will return FALSE otherwise.
}
\author{
Rob Kooper
}
