% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbitmq.R
\name{rabbitmq_get_message}
\alias{rabbitmq_get_message}
\title{Get message from RabbitMQ.}
\usage{
rabbitmq_get_message(uri, queue, count = 1, prefix = "", port = 15672)
}
\arguments{
\item{uri}{RabbitMQ URI or URL to rest endpoint}

\item{queue}{the queue the message is received from.}

\item{count}{the number of messages to retrieve from the queue.}

\item{prefix}{prefix for the rabbitmq api endpoint, default is for no prefix.}

\item{port}{port for the management interface, the default is 15672.}
}
\value{
NA if no message was retrieved, or a list of the messages payload.
}
\description{
This will get a message from RabbitMQ, if the queue does not exist it will
be created. The message will be converted to a json message that is returned.
}
\author{
Alexey Shiklomanov, Rob Kooper
}
