% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbitmq.R
\name{rabbitmq_parse_uri}
\alias{rabbitmq_parse_uri}
\title{parse the RabbiMQ URI. This will parse the uri into smaller pieces that can
be used to talk to the rest endpoint for RabbitMQ.}
\usage{
rabbitmq_parse_uri(uri, prefix = "", port = 15672)
}
\arguments{
\item{uri}{the amqp URI}

\item{prefix}{the prefix that the RabbitMQ managmenet interface uses}

\item{port}{the port for rabbitmq managment interface}
}
\value{
a list that contains the url to the mangement interface, username
password and vhost.
}
\description{
parse the RabbiMQ URI. This will parse the uri into smaller pieces that can
be used to talk to the rest endpoint for RabbitMQ.
}
