% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbitmq.R
\name{rabbitmq_post_message}
\alias{rabbitmq_post_message}
\title{Post message to RabbitMQ. This will submit a message to RabbitMQ, if the
queue does not exist it will be created. The message will be converted to
a json message that is submitted.}
\usage{
rabbitmq_post_message(uri, queue, message, prefix = "", port = 15672)
}
\arguments{
\item{uri}{RabbitMQ URI or URL to rest endpoint}

\item{queue}{the queue the message is submitted to}

\item{message}{the message to submit, will beconverted to json.}

\item{prefix}{prefix for the rabbitmq api endpoint, default is for no prefix.}

\item{port}{port for the management interface, the default is 15672.}
}
\value{
the result of the post if message was send, or NA if it failed.
}
\description{
Post message to RabbitMQ. This will submit a message to RabbitMQ, if the
queue does not exist it will be created. The message will be converted to
a json message that is submitted.
}
\author{
Alexey Shiklomanov, Rob Kooper
}
