% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-package.R
\docType{package}
\name{remote-package}
\alias{remote-package}
\alias{remote}
\title{R EMpirical Orthogonal TEleconnections}
\description{
A collection of functions to facilitate empirical orthogonal teleconnection
analysis. Some handy functions for preprocessing, such as deseasoning,
denoising, lagging are readily available for ease of usage.
}
\details{
R EMpirical Orthogonal TEleconnections
}
\references{
Empirical Orthogonal Teleconnections\cr
H. M. van den Dool, S. Saha, A. Johansson (2000)\cr
Journal of Climate, Volume 13, Issue 8 (April 2000) pp. 1421 - 1435\cr

Empirical methods in short-term climate prediction\cr
H. M. van den Dool (2007)\cr
Oxford University Press, Oxford, New York (2007)\cr
}
\seealso{
\pkg{remote} is built upon the \pkg{terra} package, which is the
direct successor of the \pkg{raster} package. The \pkg{terra} package is used
for raster data handling and processing, while \pkg{raster} is still
supported but not actively developed anymore. For more information on raster
data handling, please refer to the documentation of the \pkg{terra} package
and its functions.
}
\author{
Tim Appelhans, Florian Detsch, Thomas Nauss\cr
\cr
\emph{Maintainer:} Tim Appelhans \email{tim.appelhans@gmail.com}
}
\keyword{package}
