% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.copy.to.R
\name{remote.copy.to}
\alias{remote.copy.to}
\title{Copy file/dir to remote server from local server}
\usage{
remote.copy.to(host, src, dst, options = NULL, delete = FALSE,
  stderr = FALSE)
}
\arguments{
\item{host}{host structure to execute command on}

\item{src}{local file/dir to copy}

\item{dst}{remote file/dir to copy to}

\item{options}{additional arguments to be passed to rsync command}

\item{delete}{in case of local dir should all non-existent files be removed}

\item{stderr}{should stderr be returned as well.}
}
\value{
output of command executed
}
\description{
Copies the file/dir to the remote server from the local server. If the dst
is a folder it will copy the file into that folder.
}
\examples{
\dontrun{
  host <- list(name='geo.bu.edu', user='kooper', tunnel='/tmp/geo.tunnel')
  remote.copy.to(host, '/tmp/kooper', '/tmp/kooper', delete=TRUE)
}
}
\author{
Rob Kooper
}
