% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.execute.cmd.R
\name{remote.execute.cmd}
\alias{remote.execute.cmd}
\title{Execute command remotely}
\usage{
remote.execute.cmd(host, cmd, args = character(), stderr = FALSE)
}
\arguments{
\item{host}{host structure to execute command on}

\item{cmd}{the system command to be invoked, as a character string.}

\item{args}{a character vector of arguments to command.}

\item{stderr}{should stderr be returned as well.}
}
\value{
the captured output of the command (both stdout and stderr)
}
\description{
Execute command remotely
}
\details{
Executes the given command on the remote host using ssh. If the user is set
the system will login as the given user. If the host given is the local
machine it will execute the command locally without ssh.
}
\examples{
\dontrun{
host <- list(name = "geo.bu.edu", user = "kooper", tunnel = "/tmp/geo.tunnel")
print(remote.execute.cmd(host, "ls", c("-l", "/"), stderr = TRUE))
}
}
\author{
Rob Kooper
}
