% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_modellauncher.R
\name{setup_modellauncher}
\alias{setup_modellauncher}
\title{Setup model launcher script and job list}
\usage{
setup_modellauncher(run, rundir, host_rundir, mpirun, binary)
}
\arguments{
\item{run}{(numeric) run ID, as an integer}

\item{rundir}{Local run directory. Usually from \code{settings$rundir}}

\item{host_rundir}{Remote host run directory. Usually from \code{settings$host$rundir}}

\item{mpirun}{MPI info, usually from \code{settings$host$modellauncher$mpirun}}

\item{binary}{Binary info, usually from \code{settings$host$modellauncher$binary}}
}
\description{
Setup model launcher script and job list
}
