% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp_started}
\alias{stamp_started}
\alias{stamp_finished}
\title{Stamp start and stop times of runs}
\usage{
stamp_started(con, run)

stamp_finished(con, run)
}
\arguments{
\item{con}{BETY database connection}

\item{run}{(numeric) run ID}
}
\value{
\code{NULL}
}
\description{
Stamp start and stop times of runs
}
