% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_qsub.R
\name{start_qsub}
\alias{start_qsub}
\title{Start qsub runs}
\usage{
start_qsub(run, qsub_string, rundir, host, host_rundir, host_outdir,
  stdout_log, stderr_log, job_script, qsub_extra = NULL)
}
\arguments{
\item{run}{(numeric) run ID, as an integer}

\item{qsub_string}{qsub command string, with arguments. Usually from \code{settings$host$qsub}}

\item{rundir}{Local run directory. Usually from \code{settings$rundir}}

\item{host}{Remote host, as a list or character. Usually from \code{settings$host}.}

\item{host_rundir}{Remote host run directory. Usually from \code{settings$host$rundir}}

\item{host_outdir}{Remote host output directory. Usually from \code{settings$host$outdir}}

\item{stdout_log}{Logfile for redirecting \code{stdout}.}

\item{stderr_log}{Logfile for redirecting \code{stderr}}

\item{job_script}{Base name (no path) of script to run. Usually either \code{job.sh} or \code{launcher.sh}.}

\item{qsub_extra}{Extra \code{qsub} arguments. Usually from \code{settings$host$modellauncher$qsub.extra}}
}
\value{
Output of qsub command, as a character. This output can be parsed for ascertaining submission success.
}
\description{
Start qsub runs
}
