% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_serial.R
\name{start_serial}
\alias{start_serial}
\title{Start model execution in serial mode}
\usage{
start_serial(run, host, rundir, host_rundir, job_script)
}
\arguments{
\item{run}{(numeric) run ID, as an integer}

\item{host}{Remote host, as a list or character. Usually from \code{settings$host}.}

\item{rundir}{Local run directory. Usually from \code{settings$rundir}}

\item{host_rundir}{Remote host run directory. Usually from \code{settings$host$rundir}}

\item{job_script}{Base name (no path) of script to run. Usually either \code{job.sh} or \code{launcher.sh}.}
}
\value{
Output of execution command, as a character (see \code{\link[=remote.execute.cmd]{remote.execute.cmd()}}).
}
\description{
Start model execution in serial mode
}
