% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset}
\alias{subset}
\alias{subset,EotStack-method}
\alias{[[,EotStack,ANY,ANY-method}
\title{Subset modes in an \code{EotStack}}
\usage{
\S4method{subset}{EotStack}(x, subset, drop = FALSE, ...)

\S4method{[[}{EotStack,ANY,ANY}(x, i)
}
\arguments{
\item{x}{\code{EotStack} to be subset}

\item{subset}{\code{integer} or \code{character}. The modes to extract (either by
their indexes or names).}

\item{drop}{If \code{TRUE}, a single selected mode is returned as an \code{EotMode}.
Defaults to \code{FALSE}, which always returns an \code{EotStack}.}

\item{...}{Currently not used.}

\item{i}{Index(es) or name(s) to be subset, delegated to \code{\link[=subset]{subset()}}.}
}
\value{
An \code{EotMode} if a single mode is selected and \code{drop = TRUE}, otherwise an
\code{EotStack}.
}
\description{
Extract a set of modes from an \code{EotStack}.
}
\examples{
gph <- terra::unwrap(vdendool)

nh_modes <- eot(x = gph, y = NULL, n = 3, 
                standardised = FALSE, 
                verbose = TRUE)
                
subs <- subset(nh_modes, 2:3) # is the same as
subs <- nh_modes[[2:3]]

## effect of 'drop=FALSE' when selecting a single layer
subs <- subset(nh_modes, 2)
class(subs)
subs <- subset(nh_modes, 2, drop = TRUE)
class(subs)

## similarly to `drop = TRUE` above:
nh_modes[[2L]]
nh_modes[["mode_02"]]

}
