% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_remote.R
\name{test_remote}
\alias{test_remote}
\title{Test remote execution}
\usage{
test_remote(host, stderr = TRUE, ...)
}
\arguments{
\item{host}{host structure to execute command on}

\item{stderr}{should stderr be returned as well.}

\item{...}{additional arguments.}
}
\value{
\code{TRUE} is remote execution is successful.
If unsuccessful, depends on the value of \code{stderr}.
If \code{stderr} = TRUE (default), this function will throw an error.
If \code{stderr} = FALSE, this function will print a logger error and return FALSE.
}
\description{
Test remote execution
}
\examples{
# Localhost execution should always work
good_host <- list(name = "localhost")
test_remote(good_host)

bad_host <- list(name = "bigbadwolf")
if (!test_remote(bad_host, stderr = FALSE)) {
  print("Big Bad Wolf is a bad host.")
}
}
