% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.model.outputs.R
\name{transfer.model.outputs}
\alias{transfer.model.outputs}
\title{Transfer selected ecosystem model runs within PEcAn workflow from remote
assuming they are all finished}
\usage{
transfer.model.outputs(settings)
}
\arguments{
\item{settings}{pecan settings object}
}
\description{
Transfer selected ecosystem model runs within PEcAn workflow from remote
assuming they are all finished
}
\examples{
\dontrun{
  transfer.model.outputs(settings)
}
}
\author{
Félicien Meunier
}
