% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-package.R
\docType{data}
\name{vdendool}
\alias{vdendool}
\title{Mean seasonal (DJF) 700 mb geopotential heights}
\format{
A \code{PackedSpatRaster} with the following attributes:\cr

\if{html}{\out{<div class="sourceCode sh">}}\preformatted{size        : 14, 36, 50  (nrow, ncol, nlyr)
resolution  : 10, 5  (x, y)
extent      : -180, 180, 20, 90  (xmin, xmax, ymin, ymax)
coord. ref. : +proj=longlat +datum=WGS84 +no_defs
}\if{html}{\out{</div>}}
}
\source{
\url{https://psl.noaa.gov/data/gridded/data.ncep.reanalysis.derived.pressure.html}\cr
\emph{Original Source:} NOAA National Center for Environmental Prediction
}
\usage{
vdendool
}
\description{
NCEP/NCAR reanalysis data of mean seasonal (DJF) 700 mb geopotential heights from 1948 to 1998
}
\details{
NCEP/NCAR reanalysis data of mean seasonal (DJF) 700 mb geopotential heights from 1948 to 1998
}
\examples{
terra::unwrap(vdendool)
}
\references{
The NCEP/NCAR 40-year reanalysis project\cr
Kalnay et al. (1996)\cr
Bulletin of the American Meteorological Society, Volume 77, Issue 3, pp 437 - 471\cr
\doi{10.1175/1520-0477(1996)077<0437:TNYRP>2.0.CO;2}
}
