% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEot.R
\name{writeEot}
\alias{writeEot}
\alias{writeEot,EotMode-method}
\alias{writeEot,EotStack-method}
\title{Write \verb{Eot*} objects to disk}
\usage{
\S4method{writeEot}{EotMode}(x, path.out = ".", prefix = "remote", overwrite = TRUE, ...)

\S4method{writeEot}{EotStack}(x, ...)
}
\arguments{
\item{x}{An \verb{Eot*} object.}

\item{path.out}{The path to the folder to write the files to.}

\item{prefix}{A prefix to be added to the file names (see Details).}

\item{overwrite}{See \code{\link[terra:writeRaster]{terra::writeRaster()}}. In \code{\link[=writeEot]{writeEot()}}, this defaults
to \code{TRUE}.}

\item{...}{Further arguments passed to \code{\link[terra:writeRaster]{terra::writeRaster()}} for \code{EotMode}
input, or to the underlying \code{EotMode} method for \code{EotStack} input. If
'filetype' is not specified, the default \code{"RRASTER"} format (\code{.grd}) is
used.}
}
\description{
Write \verb{Eot*} objects to disk. This is merely a wrapper around
\code{\link[terra:writeRaster]{terra::writeRaster()}} so see respective help section for details.
}
\details{
\code{\link[=writeEot]{writeEot()}} will write the results of either an \code{EotMode} or an \code{EotStack}
to disk. For each mode the following files will be written:

\itemize{
\item \emph{pred_r} - the \code{SpatRaster} of the correlation coefficients
between the base point and each pixel of the predictor domain
\item \emph{pred_rsq} - as above but for the coefficient of determination
\item \emph{pred_rsq_sums} - as above but for the sums of coefficient of
determination
\item \emph{pred_int} - the \code{SpatRaster} of the intercept of the
regression equation for each pixel of the predictor domain
\item \emph{pred_slp} - same as above but for the slope of the
regression equation for each pixel of the predictor domain
\item \emph{pred_p} - the \code{SpatRaster} of the significance (p-value)
of the the regression equation for each pixel of the predictor domain
\item \emph{pred_resid} - the \code{SpatRaster} of the reduced data
for the predictor domain
}

Apart from \emph{pred_rsq_sums}, all these files are also created for
the response domain as \emph{resp_*}. These will be pasted together
with the prefix & the respective mode so that the file names will
look like, e.g.:

\emph{prefix_mode_n_pred_r.grd}

for the \code{SpatRaster} of the predictor correlation coefficient
of mode n using the standard \pkg{raster} file type (\code{.grd}).
}
\section{Functions}{
\itemize{
\item \code{writeEot(EotStack)}: EotStack

}}
\examples{
\dontrun{
gph <- terra::unwrap(vdendool)

nh_modes <- eot(x = gph, y = NULL, n = 2, 
                standardised = FALSE, 
                verbose = TRUE)

## write the complete EotStack
writeEot(nh_modes, prefix = "vdendool")

## write only one EotMode
writeEot(nh_modes[[2]], prefix = "vdendool")
}
}
\seealso{
\code{\link[terra:writeRaster]{terra::writeRaster()}}
}
