% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalize.R
\name{anomalize}
\alias{anomalize}
\title{Create an anomaly raster series}
\usage{
anomalize(x, reference = NULL, ...)
}
\arguments{
\item{x}{A \code{SpatRaster} (or \verb{Raster*}) series.}

\item{reference}{An optional single-layer \code{SpatRaster} (or \code{RasterLayer}) to
be used as the reference. Uses the overall mean of the original series if
\code{NULL} (default).}

\item{...}{If \code{reference = NULL}, additional arguments passed to
\code{\link[terra:app]{terra::app()}} when calculating the overall mean (except for 'fun').}
}
\value{
An anomaly \code{SpatRaster} series.
}
\description{
The function creates an anomaly raster series either based on
the overall mean of the original series, or a supplied reference raster.
For the creation of seasonal anomalies use \code{\link[=deseason]{deseason()}}.
}
\examples{
pcp = terra::unwrap(australiaGPCP)
pcp_anom = anomalize(pcp)

opar = par(mfrow = c(1,2))
plot(pcp[[10]], main = "original")
plot(pcp_anom[[10]], main = "anomalized")
par(opar)

}
\seealso{
\code{\link[=deseason]{deseason()}}, \code{\link[=denoise]{denoise()}}
}
