% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalize.R
\name{anomalize}
\alias{anomalize}
\alias{anomalize,RasterStackBrick-method}
\alias{anomalize,SpatRaster-method}
\title{Create an anomaly raster series}
\usage{
\S4method{anomalize}{RasterStackBrick}(x, ...)

\S4method{anomalize}{SpatRaster}(x, reference = NULL, ...)
}
\arguments{
\item{x}{A \code{SpatRaster} (or \verb{Raster*}) series.}

\item{...}{Additional arguments passed to \code{\link[terra:app]{terra::app()}} when 'reference' is
\code{NULL} (e.g. 'cores', 'filename'), or to the underlying \code{SpatRaster} method
for \verb{Raster*} input.}

\item{reference}{An optional \code{SpatRaster} (or \code{RasterLayer}) to be used as
the reference.}
}
\value{
An anomaly \code{SpatRaster} series.
}
\description{
The function creates an anomaly raster series either based on
the overall mean of the original series, or a supplied reference raster.
For the creation of seasonal anomalies use \code{\link[=deseason]{deseason()}}.
}
\examples{
pcp = terra::unwrap(australiaGPCP)
pcp_anom = anomalize(pcp)

opar = par(mfrow = c(1,2))
plot(pcp[[10]], main = "original")
plot(pcp_anom[[10]], main = "anomalized")
par(opar)
}
\seealso{
\code{\link[=deseason]{deseason()}}, \code{\link[=denoise]{denoise()}}
}
