% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVar.R
\name{calcVar}
\alias{calcVar}
\title{Calculate space-time variance of a RasterStack or RasterBrick}
\usage{
calcVar(x, standardised = FALSE, ...)
}
\arguments{
\item{x}{a RasterStack or RasterBrick}

\item{standardised}{logical.}

\item{...}{currently not used}
}
\value{
the mean (optionally standardised) space-time variance.
}
\description{
The function calculates the (optionally standardised) space-time 
variance of a RasterStack or RasterBrick.
}
\examples{
data("pacificSST")

calcVar(pacificSST)
}

