% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVar.R
\name{calcVar}
\alias{calcVar}
\title{Calculate space-time variance of a raster series}
\usage{
calcVar(x, standardised = FALSE, ...)
}
\arguments{
\item{x}{A \code{SpatRaster} (or \verb{Raster*}) series.}

\item{standardised}{\code{logical}, defaults to \code{FALSE}.}

\item{...}{Currently not used.}
}
\value{
The mean (optionally standardised) space-time variance as \code{numeric}.
}
\description{
The function calculates the (optionally standardised) space-time
variance of a raster series.
}
\examples{
sst = terra::unwrap(pacificSST)

calcVar(sst) # default non-standardised
calcVar(sst, standardised = TRUE)

}
