% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_run.R
\name{check_model_run}
\alias{check_model_run}
\title{Check if model run was successful}
\usage{
check_model_run(out, stop.on.error = TRUE)
}
\arguments{
\item{out}{Output from model execution, as a character.}

\item{stop.on.error}{Throw error if \emph{any} of the runs fails. Default TRUE.}
}
\value{
\code{TRUE} if model run succeeded. If model run failed, throw an error if \code{stop.on.error}, or return FALSE.
}
\description{
Check if model run was successful
}
