% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutStack.R
\name{cutStack}
\alias{cutStack}
\alias{cutStack,RasterStackBrick-method}
\alias{cutStack,SpatRaster-method}
\title{Shorten a raster series}
\usage{
\S4method{cutStack}{RasterStackBrick}(x, ...)

\S4method{cutStack}{SpatRaster}(x, tail = TRUE, n = NULL)
}
\arguments{
\item{x}{A \code{SpatRaster} (or \verb{Raster*}) series.}

\item{...}{Additional arguments passed to the underlying \code{SpatRaster} method.}

\item{tail}{\code{logical}. If \code{TRUE} (default) the layers will be taken off
the end of the stack. If \code{FALSE} layers will be taken off
the beginning.}

\item{n}{The number of layers to take away as \code{integer}. If \code{NULL} (default),
'x' is returned unchanged.}
}
\value{
A \code{SpatRaster} series shortened by 'n' layers either from the
beginning or the end, depending on the specification of 'tail'.
}
\description{
The function cuts a specified number of layers off a raster
series in order to create a lagged stack.
}
\examples{
pcp = terra::unwrap(australiaGPCP)

# 6 layers from the beginning
cutStack(pcp, tail = FALSE, n = 6)
# 8 layers from the end
cutStack(pcp, tail = TRUE, n = 8)

}
