% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoWeight.R
\name{geoWeight}
\alias{geoWeight}
\title{Geographic weighting}
\usage{
geoWeight(x, f = cos, ...)
}
\arguments{
\item{x, f, ...}{See \code{\link[=getWeights]{getWeights()}}.}
}
\value{
A weighted \code{SpatRaster}.
}
\description{
The function performs geographic weighting of non-projected long/lat data. By
default it uses the cosine of latitude (in radians) to compensate for the
area distortion, though the user can supply other weighting functions.
}
\examples{
gph <- terra::unwrap(vdendool)
wgtd <- geoWeight(gph)

opar <- par(mfrow = c(1,2))
plot(gph[[1]], main = "original")
plot(wgtd[[1]], main = "weighted")
par(opar)

}
