% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeights.R
\name{getWeights}
\alias{getWeights}
\title{Calculate weights from latitude}
\usage{
getWeights(x, f = cos, ...)
}
\arguments{
\item{x}{A non-projected \code{SpatRaster} (or \verb{Raster*}) object.}

\item{f}{A \code{function} applied to the latitude (in radians) to compute
weights. Defaults to \code{cos}.}

\item{...}{Additional arguments passed to 'f'.}
}
\value{
A \code{numeric} vector of weights for non-\code{NA} cells in 'x'.
}
\description{
Calculate weights using the cosine of latitude to compensate for
area distortion of non-projected lat/lon data.
}
\examples{
pcp = terra::unwrap(australiaGPCP)

wghts = getWeights(pcp)
utils::head(wghts)

wghts_rst = terra::setValues(pcp[[1]], wghts)

opar = par(mfrow = c(1,2))
plot(pcp[[1]], main = "data")
plot(wghts_rst, main = "weights")
par(opar)

}
