% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getWeights}
\alias{getWeights}
\title{Calculate weights from latitude}
\usage{
getWeights(x)
}
\arguments{
\item{x}{a Raster* object}
}
\value{
a numeric vector of weights
}
\description{
Calculate weights using the cosine of latitude to compensate for area
distortion of non-projected lat/lon data
}
\examples{
data("australiaGPCP")
wghts <- getWeights(australiaGPCP)
wghts_rst <- australiaGPCP[[1]]
wghts_rst[] <- wghts

opar <- par(mfrow = c(1,2))
plot(australiaGPCP[[1]], main = "data")
plot(wghts_rst, main = "weights")
par(opar)
}

