% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longtermMeans.R
\name{longtermMeans}
\alias{longtermMeans}
\alias{longtermMeans,RasterStackBrick-method}
\alias{longtermMeans,SpatRaster-method}
\title{Calculate long-term means from a raster series}
\usage{
\S4method{longtermMeans}{RasterStackBrick}(x, ...)

\S4method{longtermMeans}{SpatRaster}(x, cycle.window = 12L)
}
\arguments{
\item{x}{A \code{SpatRaster} (or \verb{Raster*}) series.}

\item{...}{For \verb{Raster*} input, arguments passed to the underlying
\code{SpatRaster} method.}

\item{cycle.window}{\code{integer}, defaults to \code{12}. See \code{\link[=deseason]{deseason()}}.}
}
\value{
A \code{SpatRaster} with 'cycle.window' layers, each containing the mean across
all corresponding time steps.
}
\description{
Calculate long-term means from an input raster series. Ideally, the number of
input layers should be divisable by the supplied 'cycle.window'. For
instance, if 'x' consists of monthly layers, 'cycle.window' should be a
multiple of \code{12}.
}
\examples{
pcp = terra::unwrap(australiaGPCP)

longtermMeans(pcp)

}
\seealso{
\code{\link[=deseason]{deseason()}}.
}
\author{
Florian Detsch
}
