% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{names}
\alias{names}
\alias{names,EotMode-method}
\alias{names,EotStack-method}
\title{Names of Eot* objects}
\usage{
\S4method{names}{EotStack}(x)

\S4method{names}{EotMode}(x)
}
\arguments{
\item{x}{a EotMode or EotStack}
}
\value{
if \code{x} is a EotStack, the names of the modes,
if \code{x} is a EotMode, the names of the slots
}
\description{
Names of Eot* objects
}
\details{
retrieves the names of Eot* objects
}
\examples{
data(vdendool)

nh_modes <- eot(vdendool, n = 2)

## mode names
names(nh_modes)

## slot names
names(nh_modes[[2]])
}

