% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{nmodes}
\alias{nmodes}
\title{Number of modes of an EotStack}
\usage{
\S4method{nmodes}{EotStack}(x)
}
\arguments{
\item{x}{an EotStack}
}
\value{
integer
}
\description{
Number of modes of an EotStack
}
\details{
retrieves the number of modes of an EotStack
}
\examples{
data(vdendool)

nh_modes <- eot(vdendool, n = 2)

nmodes(nh_modes)
}

