% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-package.R
\docType{data}
\name{pacificSST}
\alias{pacificSST}
\title{Monthly SSTs for the tropical Pacific Ocean}
\format{a RasterBrick with the following attributes\cr
\cr
dimensions  : 30, 140, 4200, 348  (nrow, ncol, ncell, nlayers)\cr
resolution  : 1, 1  (x, y)\cr
extent      : 150, 290, -15, 15  (xmin, xmax, ymin, ymax)\cr
coord. ref. : +proj=longlat +ellps=WGS84 +towgs84=0,0,0,0,0,0,0 +no_defs\cr}
\description{
Monthly NOAA sea surface temperatures for the tropical Pacific Ocean from 1982/01 to 2010/12
}
\details{
Monthly NOAA sea surface temperatures for the tropical Pacific Ocean from 1982/01 to 2010/12
}
\references{
Daily High-Resolution-Blended Analyses for Sea Surface Temperature\cr
Reynolds et al. (2007)\cr
Journal of Climate, Volume 20, Issue 22, pp. 5473 - 5496\cr
\url{http://dx.doi.org/10.1175/2007JCLI1824.1}
}

