% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub_parallel.R
\name{qsub_parallel}
\alias{qsub_parallel}
\title{qsub_parallel}
\usage{
qsub_parallel(
  settings,
  files = NULL,
  prefix = "sipnet.out",
  sleep = 10,
  hybrid = TRUE
)
}
\arguments{
\item{settings}{pecan settings object}

\item{files}{allow submit jobs based on job.sh file paths.}

\item{prefix}{used for detecting if jobs are completed or not.}

\item{sleep}{time (in second) that we wait each time for the jobs to be completed.}

\item{hybrid}{A Boolean argument decide the way of detecting job completion. If it's TRUE then we will detect both the outputted files and job ids on the server. If it's FALSE then we will only detect the job ids on the server.}
}
\description{
qsub_parallel
}
\examples{
\dontrun{
  qsub_parallel(settings)
}
}
\author{
Dongchen Zhang
}
