% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEot.R
\name{readEot}
\alias{readEot}
\title{Read \code{Eot}* files from disk}
\usage{
readEot(x, prefix = "remote", suffix = "grd")
}
\arguments{
\item{x}{\code{character}, search path for \code{Eot}* related files passed 
to \code{\link{list.files}}.}

\item{prefix}{\code{character}, see \code{\link{writeEot}} for details.}

\item{suffix}{\code{character}, file extension depending on the output file 
type of locally stored \code{Eot}* files, see \code{\link{writeRaster}}.}
}
\value{
An \code{Eot}* object.
}
\description{
Read \code{Eot}* related files from disk, e.g. for further use with 
\code{\link[remote]{predict}}.
}
\examples{
\dontrun{
## calculate 3 leading modes
data(vdendool)
nh_modes <- eot(x = vdendool, n = 3, standardised = FALSE, 
                write.out = TRUE, path.out = "~/data")
                
## reimport related files
rm(nh_modes)
nh_modes <- readEot("~/data")
nh_modes
}

}
\author{
Florian Detsch
}
\seealso{
\code{\link{eot}}, \code{\link{writeEot}}, 
\code{\link{writeRaster}}.
}

