% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.execute.R.R
\name{remote.execute.R}
\alias{remote.execute.R}
\title{Remotely execute R code}
\usage{
remote.execute.R(code, host, stderr = TRUE, ...)
}
\arguments{
\item{code}{Unevaluated R expression containing code to be run on remote. To generate, use the \code{quote()} function.}

\item{host}{host structure to execute command on}

\item{stderr}{should stderr be returned as well.}

\item{...}{Additional arguments passed to \code{\link[=remote.execute.cmd]{remote.execute.cmd()}}.}
}
\value{
Exactly the output of \code{code}, or \code{NULL} if no \code{dput()} statement can be found in the wrapped code.
}
\description{
Runs an unevaluated R expression remotely. Wrap R code in \code{quote({...})} to generate an unevaluated expression.
Wrapped code should contain a \code{dput()} statement, or a warning will be thrown.
\code{dput()} is used to return outputs, so if it is absent, the output will be `
}
\examples{
host <- list(name = "localhost")
code <- quote({
  x <- 5
  y <- 10
  out <- list(xx = seq_len(x), yy = seq_len(y) * 2)
  dput(out)
})
result <- remote.execute.R2(code = code, host = host)
}
\author{
Alexey Shiklomanov
}
