% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_rabbitmq.R
\name{start_rabbitmq}
\alias{start_rabbitmq}
\title{Start model execution using rabbitmq}
\usage{
start_rabbitmq(folder, rabbitmq_uri, rabbitmq_queue)
}
\value{
Output of execution command, as a character (see \code{\link[=remote.execute.cmd]{remote.execute.cmd()}}).
}
\description{
Start model execution using rabbitmq
}
