% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-package.R
\docType{data}
\name{vdendool}
\alias{vdendool}
\title{Mean seasonal (DJF) 700 mb geopotential heights}
\format{a RasterBrick with the following attributes\cr
\cr
dimensions  : 14, 36, 504, 50  (nrow, ncol, ncell, nlayers)\cr
resolution  : 10, 4.931507  (x, y)\cr
extent      : -180, 180, 20.9589, 90  (xmin, xmax, ymin, ymax)\cr
coord. ref. : +proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0\cr}
\source{
\url{http://www.esrl.noaa.gov/psd/data/gridded/data.ncep.reanalysis.derived.pressure.html}\cr
\emph{Original Source:} NOAA National Center for Environmental Prediction
}
\description{
NCEP/NCAR reanalysis data of mean seasonal (DJF) 700 mb geopotential heights from 1948 to 1998
}
\details{
NCEP/NCAR reanalysis data of mean seasonal (DJF) 700 mb geopotential heights from 1948 to 1998
}
\references{
The NCEP/NCAR 40-year reanalysis project\cr
Kalnay et al. (1996)\cr
Bulletin of the American Meteorological Society, Volume 77, Issue 3, pp 437 - 471\cr
\url{http://journals.ametsoc.org/doi/abs/10.1175/1520-0477(1996)077%3C0437%3ATNYRP%3E2.0.CO%3B2
}
}

